requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
python-dotenv==1.0.0
Flask-WTF==1.1.1
Werkzeug==2.3.7
WTForms==3.0.1
PyMySQL==1.1.0
cryptography==41.0.7
